Added SP 114
Added a 3 bit int to ArchipelagoSettings for the starting episode, which can be fetched via SP 114. A value of 0 indicates the starting episode is the main episode, meaning startWithMainGameUnlocked is now redundant and has been removed. Cryptic, please make sure to check the changes to extern.h and update your code accordingly.