hecka...
For context: Originally, the starter was calculated using all 16 characters of the slot name. The issue with this is that the hero name menu writes a null byte to the 11th character to truncate it to 10 characters. This new value (the null byte) would cause a different starter to be picked if you deleted the save file and did the test again without resetting. The planned fix was to only use the first 10 characters of the name (the ones that don't get touched), but... I believe Hecka forgot about zero-indexing, and used index 10 as well, meaning the byte that would get nullified was STILL being read. This change fixes that.
(you know, now that I think about it, we could theoretically read the 12th-16th characters without issue. We can do that for more randomness later on, but it's low priority.)
ov36: 2721792
hintables: 3299328