Rewrote name autofill to use a buffer
Turns out writing a null byte to the 11th character only worked in the short term, and it becomes really inconvenient once you start using the slot name for other stuff. So now we just copy the first 10 characters into a buffer, and pass the buffer to the name menu. This way, the slot name is no longer being modified at runtimee, fixing the underlying issues with random starters not being consistent. We're also using all 16 characters for hero randomization now that we can do so safely!
ov36: 2721792
hintables: 3299328